+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.2.4...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.2.5...HEAD)
+
+## [1.2.5](https://github.com/fivetran/go-fivetran/compare/v1.2.4...v1.2.5)
+
+## Added
+Support for custom config (`map[string]interface{}`) for destinations:
+- `DestinationSetupTestsService.DoCustom` method
 
 ## [1.2.4](https://github.com/fivetran/go-fivetran/compare/v1.2.3...v1.2.4)